requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
braceexpand==0.1.7
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
discord.py==2.3.2
frozenlist==1.4.0
humanize==4.8.0
idna==3.4
import-expression==1.1.4
importlib-metadata==6.8.0
jishaku==2.5.1
line-profiler==4.1.1
multidict==6.0.4
numpy==1.24.4
requests==2.31.0
typing_extensions==4.8.0
urllib3==2.0.6
Wand==0.6.11
yarl==1.9.2
zipp==3.17.0
